수정 IndexData
기존에 만든 DummyData를 삭제했습니다.
중복되었던 IndexDataMapper를 IndexDataMapper와 IndexDataIntegrationMapper로 나눴습니다.
그 외 IndexData의 생성,조회,수정,삭제 테스트 도중에 발생한 에러들을 수정했습니다.
