+++ src/main/java/com/codeit/findex/entity/IndexInfo.java
@@ -2,9 +2,8 @@ package com.codeit.findex.entity;
 
 import com.codeit.findex.entity.base.BaseEntity;
 import com.codeit.findex.entityEnum.SourceType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
+
 import java.time.LocalDate;
 import lombok.Getter;
 import lombok.Setter;
@@ -39,6 +38,7 @@ public class IndexInfo extends BaseEntity {
   private double baseIndex;
 
   /** 정보가 입력된 출처 (예: "USER", "OPEN_API") */
+  @Enumerated(EnumType.STRING)
   @Column(name = "source_type", length = 20, nullable = false)
   private SourceType sourceType;
