+++ src/main/java/com/codeit/findex/mapper/GenericMapper.java
@@ -8,5 +8,4 @@ public interface GenericMapper<D,E> {
 
     List<D> toDtos(List<E> entities);
     List<E> toEntities(List<D> dtos);
-
 }
